requirements.txt
django
django-ninja
django-ninja-extra
django-ninja-jwt